Introducing common FileSystem functions
