+++ components/SuggestionPlayerForm.js
@@ -2,62 +2,19 @@ import React, { useState } from 'react';
 
 import BaseForm from './BaseForm.js';
 import Loading from '../components/Loading.js'
-import * as FS from 'expo-file-system';
+import { readWorkingSave, addKeys } from '../global/FileSystem.js';
 
 export default function SuggestionPlayerForm(props) {
 
     const [playersLoaded, setPlayersLoaded] = useState(false);
-
     const [players, setPlayers] = useState(null);
 
-    function getWorkingSave() {
-        return new Promise((resolve, reject) => {
-            console.log('[BEGIN] Reading appState...');
-
-            let msgSuccess = '[SUCCESS] Success getting working save.';
-            let msgFailure = '[ERROR] Error getting working save.';
-            let uri = FS.documentDirectory + 'appState.json';
-            FS.readAsStringAsync(uri)
-                .then((contents) => {
-                    let appState = JSON.parse(contents);
-                    console.log(msgSuccess);
-                    resolve(appState.workingSave);
-                })
-                .catch(() => {
-                    console.log(msgFailure);
-                    reject();
-                });
-        });
-    }
-
     async function getPlayers() {
-        let file = await getWorkingSave();
-
-        console.log(`[BEGIN] Reading working save "${file}"...`);
-
-        let msgSuccess = 'Success';
-        let msgFailure = 'Failure';
-        let uri = FS.documentDirectory + file;
-        FS.readAsStringAsync(uri)
-            .then((contents) => {
-                let save = JSON.parse(contents);
-                let foo = addKeys(save.players);
-                setPlayers(foo);
-                setPlayersLoaded(true);
-                console.log(msgSuccess);
-            })
-            .catch(() => {
-                console.log(msgFailure);
-            });
-    }
-
-    function addKeys(arr) {
-        return arr.map((value, index) => (
-            {
-                key: index.toString(),
-                name: value
-            }
-        ));
+        let save = await readWorkingSave();
+        save = JSON.parse(save);
+        let players = addKeys(save.players);
+        setPlayers(players);
+        setPlayersLoaded(true);
     }
 
     if (playersLoaded) {
